backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "blinker>=1.9.0",
    "click>=8.3.1",
    "extended>=0.0.8",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "flask-jwt-extended",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "mako>=1.3.10",
    "markupsafe>=3.0.3",
    "pyjwt",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "typing-extensions>=4.15.0",
    "werkzeug>=3.1.5",
]

[tool.uv.workspace]
members = [
    "docket-sheet-importance",
]
